Preferences refactor
