+++ backend/preferences/__init__.py
@@ -1,0 +1,1 @@
+from .preference_func import get_addon_root, get_prefs
