+++ backend/preview_collections.py
@@ -12,7 +12,7 @@ import bpy
 
 from ..human.base.exceptions import HumGenException  # type: ignore
 from .logging import hg_log
-from .preference_func import get_prefs
+from .preferences import get_prefs
 
 # from HumGen3D.human.human import Human
