+++ backend/property_functions.py
@@ -7,7 +7,7 @@ from pathlib import Path
 
 import bpy
 from HumGen3D.backend.logging import hg_log
-from HumGen3D.backend.preference_func import get_prefs
+from HumGen3D.backend.preferences import get_prefs
 from HumGen3D.human.human import Human
 from HumGen3D.user_interface.feedback_func import ShowMessageBox  # type:ignore
