+++ backend/update.py
@@ -6,7 +6,7 @@ import json
 import bpy  # type: ignore
 import requests  # type: ignore
 from HumGen3D.backend.logging import hg_log
-from HumGen3D.backend.preference_func import get_prefs
+from HumGen3D.backend.preferences import get_prefs
 
 
 def check_update():
