+++ batch_generator/primitives.py
@@ -2,7 +2,7 @@ import os
 from pathlib import Path
 
 import bpy
-from HumGen3D.backend.preference_func import get_addon_root
+from HumGen3D.backend.preferences import get_addon_root
 from HumGen3D.human.base.collections import add_to_collection  # type:ignore
