+++ human/finalize_phase/clothing/pattern.py
@@ -2,7 +2,7 @@ import os
 from pathlib import Path
 
 import bpy
-from HumGen3D.backend.preference_func import get_prefs
+from HumGen3D.backend.preferences import get_prefs
 from HumGen3D.human.base.decorators import injected_context
 from HumGen3D.human.base.pcoll_content import PreviewCollectionContent
 from HumGen3D.user_interface.feedback_func import ShowMessageBox
