+++ human/finalize_phase/pose/pose.py
@@ -1,7 +1,7 @@
 import bpy
 from HumGen3D.backend.logging import hg_log
 from HumGen3D.backend.memory_management import hg_delete
-from HumGen3D.backend.preference_func import get_prefs
+from HumGen3D.backend.preferences import get_prefs
 from HumGen3D.human.base.decorators import injected_context
 from HumGen3D.human.base.pcoll_content import PreviewCollectionContent
