+++ human/hair/hair.py
@@ -1,7 +1,7 @@
 import os
 
 import bpy
-from HumGen3D.backend.preference_func import get_addon_root
+from HumGen3D.backend.preferences import get_addon_root
 from HumGen3D.human.base.decorators import cached_property
 from HumGen3D.human.base.prop_collection import PropCollection
 from HumGen3D.human.hair.eyelashes import EyelashSettings
