+++ human/human.py
@@ -10,7 +10,7 @@ from bpy.types import Object
 
 from ..backend.logging import hg_log
 from ..backend.memory_management import hg_delete, remove_broken_drivers
-from ..backend.preference_func import get_prefs
+from ..backend.preferences import get_prefs
 from ..backend.preview_collections import refresh_pcoll
 from .base.collections import add_to_collection
 from .base.decorators import cached_property, injected_context
