+++ human/skin/skin.py
@@ -7,7 +7,7 @@ from typing import TYPE_CHECKING, Any, List, Tuple
 
 import bpy
 from bpy.types import bpy_prop_collection
-from HumGen3D.backend.preference_func import get_prefs
+from HumGen3D.backend.preferences import get_prefs
 from HumGen3D.human.base.decorators import cached_property
 from HumGen3D.human.base.pcoll_content import PreviewCollectionContent
 from HumGen3D.user_interface.feedback_func import ShowMessageBox
