+++ tests/human/test_skin.py
@@ -1,14 +1,13 @@
 import random
 
 import pytest
+from HumGen3D.human.skin.skin import SkinSettings
 from HumGen3D.tests.fixtures import (
     creation_phase_human,
     finalize_phase_human,
     reverted_human,
 )
 from pytest_lazyfixture import lazy_fixture
-
-from human.skin.skin import SkinSettings
 
 fixture_names = ["creation_phase_human", "finalize_phase_human", "reverted_human"]
