+++ utility_section/content_saving.py
@@ -10,7 +10,7 @@ from shutil import copyfile
 import bpy
 from HumGen3D.backend.logging import hg_log
 from HumGen3D.backend.memory_management import hg_delete
-from HumGen3D.backend.preference_func import get_addon_root, get_prefs
+from HumGen3D.backend.preferences import get_addon_root, get_prefs
 from HumGen3D.backend.preview_collections import refresh_pcoll  # type: ignore
 from HumGen3D.human.base.shapekey_calculator import (
     build_distance_dict,
@@ -63,9 +63,7 @@ class Content_Saving_Operator:
                 img.file_format = "JPEG"
                 img.save()
             except RuntimeError as e:
-                show_message(
-                    self, "Thumbnail image doesn't have any image data"
-                )
+                show_message(self, "Thumbnail image doesn't have any image data")
                 print(e)
 
     def save_objects_optimized(
@@ -185,9 +183,7 @@ class Content_Saving_Operator:
         Args:
             obj (Object): obj to remove shapekeys from
         """
-        for sk in [
-            sk for sk in obj.data.shape_keys.key_blocks if sk.name != "Basis"
-        ]:
+        for sk in [sk for sk in obj.data.shape_keys.key_blocks if sk.name != "Basis"]:
             obj.shape_key_remove(sk)
         if obj.data.shape_keys:
             obj.shape_key_remove(obj.data.shape_keys.key_blocks["Basis"])
@@ -410,9 +406,7 @@ class HG_OT_SAVEPRESET(bpy.types.Operator, Content_Saving_Operator):
 
         self.thumb = self.sett.preset_thumbnail_enum
 
-        self.folder = pref.filepath + str(
-            Path(f"/models/{self.hg_rig.HG.gender}/")
-        )
+        self.folder = pref.filepath + str(Path(f"/models/{self.hg_rig.HG.gender}/"))
         self.name = self.sett.preset_name
         if os.path.isfile(str(Path(f"{self.folder}/{self.name}.json"))):
             return context.window_manager.invoke_props_dialog(self)
@@ -502,8 +496,7 @@ class HG_OT_SAVEPRESET(bpy.types.Operator, Content_Saving_Operator):
         sk_dict = {
             sk.name: sk.value
             for sk in sks
-            if sk.value != 0
-            and not sk.name.startswith(("expr", "cor", "Basis"))
+            if sk.value != 0 and not sk.name.startswith(("expr", "cor", "Basis"))
         }
         preset_data["shapekeys"] = sk_dict
 
@@ -536,9 +529,7 @@ class HG_OT_SAVEPRESET(bpy.types.Operator, Content_Saving_Operator):
 
         mat_dict["diffuse"] = img_name
         nodegroup_dict = {}
-        for node in [
-            node for node in nodes if node.bl_idname == "ShaderNodeGroup"
-        ]:
+        for node in [node for node in nodes if node.bl_idname == "ShaderNodeGroup"]:
             input_dict = {}
             for input_socket in [inp for inp in node.inputs if not inp.links]:
                 inp_value = (
@@ -568,9 +559,7 @@ class HG_OT_SAVEPRESET(bpy.types.Operator, Content_Saving_Operator):
         eye_mat = hg_eyes[0].data.materials[1]
         eye_nodes = eye_mat.node_tree.nodes
         mat_dict["eyes"] = {
-            "HG_Eye_Color": tuple(
-                eye_nodes["HG_Eye_Color"].inputs[2].default_value
-            ),
+            "HG_Eye_Color": tuple(eye_nodes["HG_Eye_Color"].inputs[2].default_value),
             "HG_Scelera_Color": tuple(
                 eye_nodes["HG_Scelera_Color"].inputs[2].default_value
             ),
@@ -597,16 +586,12 @@ class HG_OT_SAVEHAIR(bpy.types.Operator, Content_Saving_Operator):
             pass  # TODO unhide_human(self.hg_rig)
         except Exception as e:
             show_message(self, "Could not find human, did you delete it?")
-            hg_log(
-                "Content saving failed, rig could not be found with error: ", e
-            )
+            hg_log("Content saving failed, rig could not be found with error: ", e)
             return {"CANCELLED"}
 
         self.thumb = self.sett.preset_thumbnail_enum
 
-        self.folder = pref.filepath + str(
-            Path(f"/hair/{self.sett.save_hairtype}/")
-        )
+        self.folder = pref.filepath + str(Path(f"/hair/{self.sett.save_hairtype}/"))
         self.name = self.sett.hairstyle_name
         if os.path.isfile(str(Path(f"{self.folder}/{self.name}.blend"))):
             return context.window_manager.invoke_props_dialog(self)
@@ -635,9 +620,7 @@ class HG_OT_SAVEHAIR(bpy.types.Operator, Content_Saving_Operator):
         )
 
         keep_vgs = self._find_vgs_used_by_hair(hair_obj)
-        for vg in [
-            vg for vg in hair_obj.vertex_groups if vg.name not in keep_vgs
-        ]:
+        for vg in [vg for vg in hair_obj.vertex_groups if vg.name not in keep_vgs]:
             hair_obj.vertex_groups.remove(vg)
 
         genders = [
@@ -658,9 +641,7 @@ class HG_OT_SAVEHAIR(bpy.types.Operator, Content_Saving_Operator):
                     pref.filepath, "hair", hair_type, gender, "Custom"
                 )
             else:
-                folder = os.path.join(
-                    pref.filepath, "hair", hair_type, "Custom"
-                )
+                folder = os.path.join(pref.filepath, "hair", hair_type, "Custom")
 
             if not os.path.exists(folder):
                 os.makedirs(folder)
@@ -741,9 +722,7 @@ class HG_OT_SAVEHAIR(bpy.types.Operator, Content_Saving_Operator):
 
         for ps_name in remove_list:
             ps_idx = [
-                i
-                for i, ps in enumerate(obj.particle_systems)
-                if ps.name == ps_name
+                i for i, ps in enumerate(obj.particle_systems) if ps.name == ps_name
             ]
             obj.particle_systems.active_index = ps_idx[0]
             bpy.ops.object.particle_system_remove()
@@ -816,17 +795,11 @@ class HG_OT_SAVEOUTFIT(bpy.types.Operator, Content_Saving_Operator):
             )
             return {"CANCELLED"}
 
-        self.folder = os.path.join(
-            self.pref.filepath, self.sett.saveoutfit_categ
-        )
+        self.folder = os.path.join(self.pref.filepath, self.sett.saveoutfit_categ)
         self.name = self.sett.clothing_name
 
         if os.path.isfile(
-            str(
-                Path(
-                    f"{self.folder}/{self.hg_rig.HG.gender}/Custom/{self.name}.blend"
-                )
-            )
+            str(Path(f"{self.folder}/{self.hg_rig.HG.gender}/Custom/{self.name}.blend"))
         ):
             self.alert = "overwrite"
             return context.window_manager.invoke_props_dialog(self)
@@ -869,11 +842,7 @@ class HG_OT_SAVEOUTFIT(bpy.types.Operator, Content_Saving_Operator):
         for gender in genders:
             export_list = []
             backup_human = next(
-                (
-                    obj
-                    for obj in self.hg_rig.HG.backup.children
-                    if "hg_body" in obj
-                )
+                (obj for obj in self.hg_rig.HG.backup.children if "hg_body" in obj)
             )
             if gender == "male":
                 backup_human = backup_human.copy()
@@ -912,9 +881,7 @@ class HG_OT_SAVEOUTFIT(bpy.types.Operator, Content_Saving_Operator):
                     ignore_cor_sk=True,
                 )
                 human = None  # FIXME
-                human.creation_phase.length._correct_origin(
-                    context, obj, backup_human
-                )
+                human.creation_phase.length._correct_origin(context, obj, backup_human)
                 export_list.append(obj_copy)
 
             if gender == "male":
@@ -994,9 +961,7 @@ class HG_OT_SAVEOUTFIT(bpy.types.Operator, Content_Saving_Operator):
         if img_name in saved_images:
             return saved_images[img_name], saved_images
 
-        path = self.pref.filepath + str(
-            Path(f"{self.sett.saveoutfit_categ}/textures/")
-        )
+        path = self.pref.filepath + str(Path(f"{self.sett.saveoutfit_categ}/textures/"))
         if not os.path.exists(path):
             os.makedirs(path)
 
@@ -1008,9 +973,7 @@ class HG_OT_SAVEOUTFIT(bpy.types.Operator, Content_Saving_Operator):
             )
             saved_images[img_name] = full_path
         except RuntimeError as e:
-            hg_log(
-                f"failed to save {img.name} with error {e}", level="WARNING"
-            )
+            hg_log(f"failed to save {img.name} with error {e}", level="WARNING")
             self.report(
                 {"WARNING"},
                 "One or more images failed to save. See the system console for specifics",
@@ -1088,9 +1051,7 @@ class HG_OT_AUTO_RENDER_THUMB(bpy.types.Operator, Content_Saving_Operator):
             (50, (0, -1.2, 0)),
         ]
         for energy, location in light_settings_enum:
-            point_light = bpy.data.lights.new(
-                name=f"light_{energy}W", type="POINT"
-            )
+            point_light = bpy.data.lights.new(name=f"light_{energy}W", type="POINT")
             point_light.energy = energy
             point_light_object = bpy.data.objects.new("Light", point_light)
             point_light_object.location = Vector(location) + hg_rig.location
@@ -1158,9 +1119,7 @@ class HG_OT_AUTO_RENDER_THUMB(bpy.types.Operator, Content_Saving_Operator):
         type_sett = type_settings_dict[thumbnail_type]
         return type_sett
 
-    def _make_camera_look_at_human(
-        self, obj_camera, hg_rig, look_at_correction
-    ):
+    def _make_camera_look_at_human(self, obj_camera, hg_rig, look_at_correction):
         """Makes the passed camera point towards a preset point on the human
 
         Args:
@@ -1172,8 +1131,7 @@ class HG_OT_AUTO_RENDER_THUMB(bpy.types.Operator, Content_Saving_Operator):
 
         hg_loc = hg_rig.location
         height_adjustment = (
-            hg_rig.dimensions[2]
-            - look_at_correction * 0.55 * hg_rig.dimensions[2]
+            hg_rig.dimensions[2] - look_at_correction * 0.55 * hg_rig.dimensions[2]
         )
         hg_rig_loc_adjusted = Vector(
             (hg_loc[0], hg_loc[1], hg_loc[2] + height_adjustment)
